LightDM 1.26.0
LightDM bump to the latest release. Regenerated patches. Added commented
out paths to `accountsservice`, we may use it some day. Tested: title
widgets, login, restart, shutdown. Suspend does not work (but it's not a
regression for me), it just puts LightDM SMF to maintenance.

New feature: In case of joined displays, the login prompt follows
pointer to activated screen.

Bug fixed: When screen turned off in LightDM it was stuck for me and I
was not able to unlock it. Not anymore.